+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Improve help of gg dna
 - Rename »gg dna test« into »gg dna build«
 - Complain about adding non dna repos
+- Improve error messages
 
 ## 0.1.0 - 2026-08-14